requirements.txt
aiofiles==0.5.0
aiohttp==3.6.2
appdirs==1.4.4
async-timeout==3.0.1
asynctest==0.13.0
attrs==19.3.0
certifi==2020.6.20
cfgv==3.2.0
chardet==3.0.4
coverage==5.2.1
discord.py==1.4.1
distlib==0.3.1
dnspython==2.0.0
docutils==0.16
filelock==3.0.12
identify==1.4.25
idna==2.10
libneko==0.24.9
motor==2.1.0
multidict==4.7.6
nodeenv==1.4.0
pre-commit==2.6.0
pymongo==3.11.0
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
virtualenv==20.0.30
websockets==8.1
yarl==1.5.1
